Update TestPage.php
added the delay slider from 1 to 1000 and set the value to 1 (it wanted to start at 501 for whatever reason. I am not sure if we want to hold the post value from when they submit the form or just leave as is.

BONUS ROUND: added the two check buttons but left out the value as I was unsure if it needed to be OFF of True/False.

Hope these changes are up to your snuff!