updates
I think I found the code for the checkboxes, but I might be misnaming things I may need some help.

Also I have the modelbox, which was a copy paste, but it is acting weird and shutting itself down, not sure if that is a cache issue on my end.