Update TestPage.php
I found out that because the button was in a form it was somehow closing itself down thinking it was trying to possibly submit. I have moved the edit shows modal button outside the form, which appears to fix the issue of the pop up just randomly deciding to to close down on us. Not sure if this is a way we can make it work or if we possibly have to find a fix to make it work within the form