start of the model
I know we are going to discuss it to find a way to make this work as you invisioned, but I did add in at least the playlist selecter for right now until we can solve the issue. I also added the gradient color to the model box so it wasn't white on white.