Clean up and added your drop down
Using the technique we found out about yesterday with writing code once in php and including whenever we need, I used that with our Nav file. This allowed me to remove 32 lines of code to call forth HTML script.

I then went ahead and used the same method to our basic footer we have now, which I invision eventually becoming more extensive down the line maybe.

This was a huge clean up of the code and I think we will both find that locating stuff is going to be 100 times easier.