nav
I had moved testNav into Nav, but it looked as if I managed to break loging in somehow someway. It would show you logged in but if you went to another page it would somhow log you out and take you to the registration page. I left the new nav code in testNav page and included that file into index.php so we can solve that problem later.