Update ConfigForm.php
dang that was a lot of lines to remove. I kept the code in for how we are displaying the motion and light sensor and time of day not sure if we want to do a new page or just insert the features into the new tables while leaving it in one spot.