not going to lie I struggled
I won't lie I did struggle a bit with the way you told me to handle it that I just ended up doing it the way I did it. I researched that you can set a session to somethinbg so basically I said if the checkbox wasn't empty assign the session variable to 1. I tested this on the local side by echoing out works and checked in the database to made sure it returned the right varaible.