attempting to wire into edit shows
I am finding it hard to test it on local as it seems to want to let you create a new playlist, but it won't let you store/save a show into that playlist. I am not sure if Config form is a factor in that not quite working the way it should.