Config Sanity Check 1.6
I have it now writting to both databases, but for whatever reason it is wanting to pull in a zero for lightsystem id when creating a new one. I could be very well doing it wrong with inserting the ID in, but I based it off the code we had for inserting things into a features table. I am going to continue solidering on and ask you later what little bitty issue I have left to solve for this damn systemId.