started working while you were on the phone
I am not sure if this is where you want it to live on the page, but I tossed it up on show designer for the time being and started the first bit of wiring. It didn't have a matrix type that I could find, but once I know that and the parms needed, it should be quick to finish.