setting up for your successful work today
I was able to wire maybe one through the light system wiring looked like it required javascript maping to have been done, but we do have that code in lightShows.php. I renamed registration to users throughout the nav bars as well. I had time to add in the buttons too for you.